Split `IdleReason` enum from `Common.h`
