+++ OPHD/EnumIdleReason.h
@@ -1,0 +1,17 @@
+#pragma once
+
+
+enum class IdleReason
+{
+	None,
+
+	PlayerSet,
+	InternalStorageFull,
+	FactoryProductionComplete,
+	FactoryInsufficientResources,
+	FactoryInsufficientRobotCommandCapacity,
+	FactoryInsufficientWarehouseSpace,
+	MineExhausted,
+	MineInactive,
+	InsufficientLuxuryProduct
+};
